+++ react/src/blueDarkTheme.ts
@@ -20,7 +20,7 @@ const ThemeColors = {
     error: createSimplePalette(PXBColors.red),
     success: createSimplePalette(PXBColors.green),
     info: createSimplePalette(PXBColors.lightBlue),
-    divider: Color(PXBColors.black[200]).alpha(0.32).rgb().string(),
+    divider: Color(PXBColors.black[200]).alpha(0.32).string(),
     warning: {
         light: PXBColors.yellow[100],
         main: PXBColors.yellow[300],
@@ -33,14 +33,14 @@ const ThemeColors = {
     text: {
         primary: PXBColors.black[50],
         secondary: PXBColors.black[200],
-        disabled: Color(PXBColors.black[300]).alpha(0.32).rgb().string(),
-        hint: Color(PXBColors.black[300]).alpha(0.32).rgb().string(),
+        disabled: Color(PXBColors.black[300]).alpha(0.36).string(),
+        hint: Color(PXBColors.black[300]).alpha(0.36).string(),
     },
     action: {
-        hover: Color(PXBColors.black[50]).alpha(0.1).rgb().string(),
+        hover: Color(PXBColors.black[50]).alpha(0.1).string(),
         active: PXBColors.black[200],
-        disabled: Color(PXBColors.black[50]).alpha(0.1).rgb().string(),
-        disabledBackground: Color(PXBColors.black[200]).alpha(0.24).rgb().string(),
+        disabled: Color(PXBColors.black[50]).alpha(0.1).string(),
+        disabledBackground: Color(PXBColors.black[200]).alpha(0.24).string(),
     },
 };
 const WhiteText = PXBColors.white[50];
@@ -89,7 +89,7 @@ export const blueDarkTheme: ThemeOptions = {
         // AVATAR OVERRIDES
         MuiAvatar: {
             colorDefault: {
-                backgroundColor: Color(PXBColors.black[50]).alpha(0.1).rgb().string(),
+                backgroundColor: Color(PXBColors.black[50]).alpha(0.1).string(),
                 color: ThemeColors.text.primary,
             },
         },
@@ -97,14 +97,22 @@ export const blueDarkTheme: ThemeOptions = {
         // BACKDROP OVERRIDES
         MuiBackdrop: {
             root: {
-                backgroundColor: Color(PXBColors.darkBlack[900]).alpha(0.7).rgb().string(),
+                backgroundColor: Color(PXBColors.darkBlack[900]).alpha(0.7).string(),
             },
         },
 
         // BADGE OVERRIDES
         MuiBadge: {
             colorError: {
-                backgroundColor: ThemeColors.error.main,
+                backgroundColor: ThemeColors.error.dark,
+                color: WhiteText,
+            },
+            colorPrimary: {
+                backgroundColor: ThemeColors.primary.dark,
+                color: WhiteText,
+            },
+            colorSecondary: {
+                backgroundColor: ThemeColors.secondary.dark,
                 color: WhiteText,
             },
         },
@@ -132,9 +140,37 @@ export const blueDarkTheme: ThemeOptions = {
         MuiButton: {
             root: {
                 textTransform: 'none',
+                '&:hover': {
+                    backgroundColor: PXBColors.black[400],
+                },
             },
             outlined: {
-                borderColor: PXBColors.black[100],
+                borderColor: PXBColors.black[200],
+                '&$disabled': {
+                    borderColor: Color(PXBColors.black[200]).alpha(0.36).string(),
+                    color: Color(PXBColors.black[300]).alpha(0.36).string(),
+                },
+            },
+            outlinedPrimary: {
+                borderColor: ThemeColors.primary.main,
+                '&:hover': {
+                    backgroundColor: Color(ThemeColors.primary.dark).alpha(0.2).string(),
+                },
+            },
+            outlinedSecondary: {
+                '&:not($disabled)': {
+                    borderColor: ThemeColors.secondary.main,
+                    // backgroundColor: Color(ThemeColors.primary.dark).alpha(0.2).string(),
+                    '&:hover': {
+                        // color: PXBColors.blue[100],
+                        // borderColor: PXBColors.blue[100],
+                        backgroundColor: Color(ThemeColors.secondary.dark).alpha(0.2).string(),
+                    },
+                },
+                '&$disabled': {
+                    borderColor: Color(PXBColors.black[200]).alpha(0.36).string(),
+                    color: Color(PXBColors.black[300]).alpha(0.36).string(),
+                },
             },
             contained: {
                 backgroundColor: MediumBlackBackground,
@@ -142,15 +178,41 @@ export const blueDarkTheme: ThemeOptions = {
                 '&:hover': {
                     backgroundColor: PXBColors.black[300],
                 },
+                '&$disabled': {
+                    backgroundColor: Color(PXBColors.black[200]).alpha(0.24).string(),
+                    color: PXBColors.black[400],
+                },
             },
             containedPrimary: {
                 backgroundColor: ThemeColors.primary.dark,
                 color: WhiteText,
                 '&:hover': {
-                    backgroundColor: PXBColors.blue[700],
+                    backgroundColor: PXBColors.blue[300],
                 },
+            },
+            containedSecondary: {
+                backgroundColor: ThemeColors.secondary.dark,
+                color: WhiteText,
+                '&:hover': {
+                    backgroundColor: PXBColors.lightBlue[300],
+                },
+            },
+            text: {
                 '&$disabled': {
-                    color: Color(PXBColors.black[200]).alpha(0.24).rgb().string(),
+                    color: Color(PXBColors.black[300]).alpha(0.36).string(),
+                },
+                '&:hover': {
+                    backgroundColor: Color(PXBColors.black[50]).alpha(0.1).string(),
+                },
+            },
+            textPrimary: {
+                '&:hover': {
+                    backgroundColor: Color(ThemeColors.primary.dark).alpha(0.2).string(),
+                },
+            },
+            textSecondary: {
+                '&:hover': {
+                    backgroundColor: Color(ThemeColors.secondary.dark).alpha(0.2).string(),
                 },
             },
             disabled: {},
@@ -163,7 +225,16 @@ export const blueDarkTheme: ThemeOptions = {
                 backgroundColor: MediumBlackBackground,
                 color: ThemeColors.text.primary,
                 '& $avatar': {
-                    color: WhiteText,
+                    backgroundColor: PXBColors.black[700],
+                    color: ThemeColors.text.primary,
+                },
+                '& $avatarColorPrimary': {
+                    backgroundColor: ThemeColors.primary.light,
+                    color: ThemeColors.primary.dark,
+                },
+                '& $avatarColorSecondary': {
+                    backgroundColor: ThemeColors.primary.light,
+                    color: ThemeColors.primary.dark,
                 },
             },
             clickable: {
@@ -172,6 +243,7 @@ export const blueDarkTheme: ThemeOptions = {
                 },
             },
             deleteIcon: {
+                fontSize: '1.125rem',
                 color: ThemeColors.text.secondary,
                 '&:hover': {
                     color: ThemeColors.text.primary,
@@ -198,16 +270,31 @@ export const blueDarkTheme: ThemeOptions = {
             },
             outlined: {
                 backgroundColor: ThemeColors.background.paper,
-                borderColor: Color(PXBColors.black[200]).alpha(0.32).rgb().string(),
+                borderColor: Color(PXBColors.black[200]).alpha(0.32).string(),
                 '&$clickable:hover': {
                     backgroundColor: PXBColors.black[800],
                 },
+                '& $avatar': {
+                    backgroundColor: PXBColors.black[600],
+                    color: ThemeColors.text.primary,
+                },
+                '& $avatarColorPrimary': {
+                    backgroundColor: PXBColors.blue[100],
+                    color: ThemeColors.primary.dark,
+                },
+                '& $avatarColorSecondary': {
+                    backgroundColor: PXBColors.blue[100],
+                    color: ThemeColors.primary.dark,
+                },
             },
             outlinedPrimary: {
-                backgroundColor: Color(ThemeColors.primary.dark).alpha(0.2).rgb().string(),
+                backgroundColor: Color(ThemeColors.primary.dark).alpha(0.2).string(),
                 '&$clickable:hover': {
-                    backgroundColor: Color(ThemeColors.primary.dark).alpha(0.3).rgb().string(),
+                    backgroundColor: Color(ThemeColors.primary.dark).alpha(0.3).string(),
                 },
+            },
+            icon: {
+                fontSize: '1.125rem',
             },
             avatar: {},
         },
@@ -288,6 +375,9 @@ export const blueDarkTheme: ThemeOptions = {
                 '& .MuiButton-textPrimary': {
                     color: ThemeColors.primary.dark,
                 },
+                '& .MuiButton-textSecondary': {
+                    color: PXBColors.lightBlue[500],
+                },
             },
         },
 
@@ -316,7 +406,7 @@ export const blueDarkTheme: ThemeOptions = {
         },
         MuiStepIcon: {
             root: {
-                color: Color(PXBColors.black[300]).alpha(0.32).rgb().string(),
+                color: Color(PXBColors.black[300]).alpha(0.32).string(),
                 zIndex: 1,
                 '&$active': {
                     color: ThemeColors.primary.dark,
@@ -389,27 +479,20 @@ export const blueDarkTheme: ThemeOptions = {
                 color: ThemeColors.text.primary,
                 backgroundColor: PXBColors.darkBlack[300],
                 '&$hover:hover': {
-                    backgroundColor: Color(PXBColors.darkBlack[300])
-                        .mix(Color(PXBColors.black[500]), 0.5)
-                        .rgb()
-                        .string(),
+                    backgroundColor: Color(PXBColors.darkBlack[300]).mix(Color(PXBColors.black[500]), 0.5).string(),
                 },
                 '&:nth-of-type(odd):not($selected)': {
                     backgroundColor: PXBColors.black[900],
                     '&$hover:hover': {
-                        backgroundColor: Color(PXBColors.black[900])
-                            .mix(Color(PXBColors.black[500]), 0.5)
-                            .rgb()
-                            .string(),
+                        backgroundColor: Color(PXBColors.black[900]).mix(Color(PXBColors.black[500]), 0.5).string(),
                     },
                 },
                 '&$selected': {
-                    backgroundColor: Color(ThemeColors.primary.dark).alpha(0.2).rgb().string(),
+                    backgroundColor: Color(ThemeColors.primary.dark).alpha(0.2).string(),
                     '&$hover:hover': {
                         backgroundColor: Color(ThemeColors.primary.dark)
                             .mix(Color(PXBColors.black[500]), 0.5)
                             .alpha(0.2)
-                            .rgb()
                             .string(),
                     },
                 },
@@ -433,6 +516,12 @@ export const blueDarkTheme: ThemeOptions = {
 
         // TABS OVERRIDES
         MuiTab: {
+            root: {
+                fontWeight: 400,
+                '&$selected': {
+                    fontWeight: 600,
+                },
+            },
             textColorInherit: {
                 color: ThemeColors.text.secondary,
                 opacity: 1,
@@ -452,13 +541,24 @@ export const blueDarkTheme: ThemeOptions = {
         },
 
         // TEXT FIELD OVERRIDES
+        MuiInputBase: {
+            input: {
+                '&::placeholder': {
+                    color: PXBColors.black[300],
+                    opacity: 0.36,
+                },
+            },
+        },
         MuiInput: {
             underline: {
                 '&:before': {
-                    borderBottomColor: Color(ThemeColors.action.active).alpha(0.36).rgb().string(),
+                    borderBottomColor: Color(ThemeColors.action.active).alpha(0.36).string(),
                 },
                 '&:after': {
-                    borderBottomColor: ThemeColors.primary.dark, // Color(ThemeColors.action.active).alpha(.36),
+                    borderBottomColor: ThemeColors.primary.dark,
+                },
+                '&$error$focused:after': {
+                    borderBottomColor: ThemeColors.error.dark,
                 },
                 '&$colorSecondary:not($error):after': {
                     borderBottomColor: ThemeColors.secondary.main,
@@ -469,6 +569,7 @@ export const blueDarkTheme: ThemeOptions = {
                 },
             },
             disabled: {},
+            focused: {},
             colorSecondary: {},
             error: {},
         },
@@ -482,8 +583,8 @@ export const blueDarkTheme: ThemeOptions = {
                     backgroundColor: PXBColors.black[800],
                 },
                 '&$disabled': {
-                    color: Color(PXBColors.black[300]).alpha(0.36).rgb().string(),
-                    backgroundColor: Color(PXBColors.black[800]).alpha(0.5),
+                    color: Color(PXBColors.black[300]).alpha(0.36).string(),
+                    backgroundColor: Color(PXBColors.black[800]).alpha(0.5).string(),
                     pointerEvents: 'none',
                 },
             },
@@ -493,6 +594,9 @@ export const blueDarkTheme: ThemeOptions = {
                 },
                 '&:after': {
                     borderBottomColor: ThemeColors.primary.dark,
+                },
+                '&$error$focused:after': {
+                    borderBottomColor: ThemeColors.error.dark,
                 },
                 '&$disabled:before': {
                     borderBottomStyle: 'solid',
@@ -512,12 +616,15 @@ export const blueDarkTheme: ThemeOptions = {
                     borderColor: ThemeColors.error.dark,
                 },
                 '&$disabled $notchedOutline': {
-                    borderColor: Color(ThemeColors.action.active).alpha(0.36).rgb().string(),
+                    borderColor: Color(ThemeColors.action.active).alpha(0.36).string(),
+                },
+                '&:hover $notchedOutline': {
+                    borderColor: PXBColors.black[200],
                 },
                 '&$error$colorSecondary$focused $notchedOutline': {
                     borderColor: ThemeColors.error.dark,
                 },
-                '&$error:hover $notchedOutline': {
+                '&$error:hover:not($focused) $notchedOutline': {
                     borderColor: ThemeColors.error.main,
                 },
                 '&$focused $notchedOutline': {
@@ -534,7 +641,7 @@ export const blueDarkTheme: ThemeOptions = {
         },
         MuiFormLabel: {
             root: {
-                color: ThemeColors.text.secondary,
+                color: ThemeColors.text.primary,
                 '&$error$colorSecondary$focused': {
                     color: ThemeColors.error.main,
                 },
@@ -565,9 +672,9 @@ export const blueDarkTheme: ThemeOptions = {
             root: {
                 backgroundColor: ThemeColors.background.paper,
                 color: PXBColors.gray[500],
-                borderColor: Color(PXBColors.black[200]).alpha(0.32).rgb().string(),
+                borderColor: Color(PXBColors.black[200]).alpha(0.32).string(),
                 '&$selected': {
-                    backgroundColor: Color(ThemeColors.primary.dark).alpha(0.2).rgb().string(),
+                    backgroundColor: Color(ThemeColors.primary.dark).alpha(0.2).string(),
                     color: ThemeColors.primary.main,
                 },
             },