+++ react/src/shared.ts
@@ -15,6 +15,9 @@ export const typography = {
     h6: {
         fontWeight: 600,
     },
+    subtitle1: {
+        fontWeight: 600,
+    },
     subtitle2: {
         fontWeight: 600,
     },