+++ react/CHANGELOG.md
@@ -7,6 +7,7 @@
 -   Update deprecated `typeface-open-sans` dependency to `@fontsource/open-sans`.
 -   Updates to dark theme colors.
 -   Minor updates to light theme colors.
+-   `<Typography>` `subtitle1` variant from regular to semibold to distinguish it from `body1`.
 
 ## v5.0.3